refactor: extract WC leaderboard computation into service (no behavior change)
