+++ backend/src/routes/worldCupPicks.js
@@ -4,6 +4,7 @@ import { Group } from '../models/Group.js';
 import { GameService } from '../services/GameService.js';
 import { UserPick, WORLD_CUP_RESULTS } from '../models/UserPick.js';
 import { buildLeaderboard } from '../services/SoccerScoringService.js';
+import { buildGroupLeaderboard } from '../services/WorldCupLeaderboardService.js';
 import pool from '../config/database.js';
 
 const router = express.Router();
@@ -188,71 +189,14 @@ router.get('/group/:groupId/world-cup/leaderboard', authenticateToken, async (re
     const { groupId } = req.params;
     const group = await ensureMembership(groupId, req.user.id);
 
-    // Resolved games for every stage. winnerTeamId is grafted on by GameService
-    // and is the only signal of the advancing side on a PK shootout. The stages
-    // are independent, so fetch them concurrently — sequentially this was seven
-    // back-to-back round trips and dominated the endpoint's latency.
-    const stageSlates = await Promise.all(
-      WORLD_CUP_STAGES.map((stage) => GameService.getWorldCupStage(stage))
-    );
-    const games = stageSlates.flat();
-    const gameById = new Map(games.map(g => [g.id, g]));
-    const wcGameIds = [...gameById.keys()];
-
-    const { rows: members } = await pool.query(
-      `SELECT u.id, u.name, u.picture_url
-         FROM group_memberships gm
-         JOIN users u ON u.id = gm.user_id
-        WHERE gm.group_id = $1`,
-      [group.id]
-    );
-
-    // World Cup picks only (picked_result NOT NULL) for games in this tournament.
-    let pickRows = [];
-    if (wcGameIds.length > 0) {
-      const result = await pool.query(
-        `SELECT user_id, game_id, picked_result
-           FROM user_picks
-          WHERE group_id = $1 AND picked_result IS NOT NULL AND game_id = ANY($2::int[])`,
-        [group.id, wcGameIds]
-      );
-      pickRows = result.rows;
+    // Resolved games for every stage (unchanged live behavior).
+    const games = [];
+    for (const stage of WORLD_CUP_STAGES) {
+      const stageGames = await GameService.getWorldCupStage(stage);
+      games.push(...stageGames);
     }
 
-    // Shape rows for the scoring service: one (userId, pick, game) per stored pick.
-    const scoringRows = [];
-    const usersWithPicks = new Set();
-    for (const pr of pickRows) {
-      const game = gameById.get(pr.game_id);
-      if (!game) continue;
-      scoringRows.push({ userId: pr.user_id, pick: { picked_result: pr.picked_result }, game });
-      usersWithPicks.add(pr.user_id);
-    }
-    // Members without any pick still belong on the board with a zero row. A null
-    // pick scores nothing in aggregateUserScore, so it yields a clean zero.
-    for (const m of members) {
-      if (!usersWithPicks.has(m.id)) {
-        scoringRows.push({ userId: m.id, pick: null, game: null });
-      }
-    }
-
-    const memberById = new Map(members.map(m => [m.id, m]));
-    const leaderboard = buildLeaderboard(scoringRows).map(row => {
-      const u = memberById.get(row.userId) || {};
-      return {
-        userId: row.userId,
-        name: u.name ?? null,
-        pictureUrl: u.picture_url ?? null,
-        rank: row.rank,
-        tied: row.tied,
-        points: row.points,
-        wins_correct: row.wins_correct,
-        losses: row.losses,
-        draws_correct: row.draws_correct,
-        draws_incorrect: row.draws_incorrect,
-      };
-    });
-
+    const leaderboard = await buildGroupLeaderboard(pool, group, games);
     res.json({ leaderboard });
   } catch (e) {
     if (e.message === 'GROUP_NOT_FOUND') return res.status(404).json({ error: 'Group not found' });
